Convert hash table to use 64 bit keys instead of 32 bit keys, since they are used to store pointers.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2661 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/gamelib/animobj.c
@@ -61,7 +61,7 @@ static	ANIMOBJDIEDTESTFUNC g_pDiedFunc = NULL;
 /***************************************************************************/
 /* local functions */
 
-static UDWORD		animObj_HashFunction( int iKey1, int iKey2 );
+static UDWORD		animObj_HashFunction(int64_t iKey1, int64_t iKey2);
 static void		animObj_HashFreeElementFunc( void * psElement );
 
 /***************************************************************************/
@@ -119,7 +119,7 @@ animObj_SetDoneFunc( ANIM_OBJECT *psObj, ANIMOBJDONEFUNC pDoneFunc )
 /***************************************************************************/
 
 static UDWORD
-animObj_HashFunction( int iKey1, int iKey2 )
+animObj_HashFunction(int64_t iKey1, int64_t iKey2)
 {
 	return (iKey1 + iKey2)%ANIM_HASH_TABLE_SIZE;
 }
@@ -179,7 +179,7 @@ animObj_Update( void )
 		if ( bRemove == TRUE )
 		{
 			if ( hashTable_RemoveElement( g_pAnimObjTable, psObj,
-				(int) psObj->psParent, psObj->psAnim->uwID ) == FALSE )
+				(int64_t) psObj->psParent, psObj->psAnim->uwID ) == FALSE )
 			{
 				debug( LOG_ERROR, "animObj_Update: couldn't remove anim obj\n" );
 				abort();
@@ -252,7 +252,7 @@ animObj_Add( void *pParentObj, int iAnimID,
 	}
 
 	/* insert object in table by parent */
-	hashTable_InsertElement(g_pAnimObjTable, psObj, (int) pParentObj, iAnimID);
+	hashTable_InsertElement(g_pAnimObjTable, psObj, (int64_t) pParentObj, iAnimID);
 
 	return psObj;
 }
@@ -283,8 +283,7 @@ animObj_GetNext( void )
 ANIM_OBJECT *
 animObj_Find( void *pParentObj, int iAnimID )
 {
-	return (ANIM_OBJECT*)hashTable_FindElement( g_pAnimObjTable,
-										(int) pParentObj, iAnimID );
+	return (ANIM_OBJECT*)hashTable_FindElement(g_pAnimObjTable, (int64_t) pParentObj, iAnimID);
 }
 
 /***************************************************************************/
@@ -292,8 +291,7 @@ animObj_Find( void *pParentObj, int iAnimID )
 BOOL
 animObj_Remove( ANIM_OBJECT **ppsObj, int iAnimID )
 {
-	BOOL bRemOK = hashTable_RemoveElement( g_pAnimObjTable, *ppsObj,
-										(int) (*ppsObj)->psParent, iAnimID );
+	BOOL bRemOK = hashTable_RemoveElement(g_pAnimObjTable, *ppsObj, (int64_t)(*ppsObj)->psParent, iAnimID);
     //init the animation
     *ppsObj = NULL;
